renamed function in utils folder for consistency
