+++ src/utils/GoToLoginScreenFunction.ts
